Add files via upload
This task contains 232 items: 113 words, 113 non-words, 6 verbs which serve as cues. 
The words are presented for 0.5 s and separated by a cross for 0.725 s. 
Participants have to decide whether a series of letters appearing on the screen makes up a word or not.  At the same time, they have to remember to press the 'y' button every time the letters they are seeing make up a word. 
The task requires psychtoolbox installed on Matlab. 
Combine the lexical decision task practice trial with this one to have the whole prospective memory task.